Reorganized files and reversed the gradient
